requirements.txt
aiomysql
asyncpg
requests
fastapi
uvicorn
python-dotenv
pydantic-settings
pip-tools
email-validator
sqlalchemy>=1.4